megaplan: outcome-readme-docs-and-cli-20260602-2209 execute
