+++ pyproject.toml
@@ -26,7 +26,7 @@ dependencies = [
 ]
 
 [project.scripts]
-astrid = "astrid.pipeline:main"
+astrid = "astrid.gateway:main"
 
 [project.optional-dependencies]
 dev = [
